github_loader/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Deprecated==1.2.6
Flask==1.1.1
Flask-Login==0.4.1
flask-paginate==0.5.3
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
hurry.filesize==0.9
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyGithub==1.43.8
PyJWT==1.7.1
requests==2.22.0
SQLAlchemy==1.3.6
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
